+++ sentry-android-core/src/main/java/io/sentry/android/core/ApplicationNotResponding.java
@@ -4,6 +4,7 @@ package io.sentry.android.core;
 
 import io.sentry.util.Objects;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * Error thrown by ANRWatchDog when an ANR is detected. Contains the stack trace of the frozen UI
@@ -16,15 +17,15 @@ import org.jetbrains.annotations.NotNull;
 final class ApplicationNotResponding extends RuntimeException {
   private static final long serialVersionUID = 252541144579117016L;
 
-  private final Thread thread;
+  private final @NotNull Thread thread;
 
-  ApplicationNotResponding(@NotNull String message, @NotNull Thread thread) {
+  ApplicationNotResponding(final @Nullable String message, final @NotNull Thread thread) {
     super(message);
     this.thread = Objects.requireNonNull(thread, "Thread must be provided.");
     setStackTrace(this.thread.getStackTrace());
   }
 
-  public Thread getThread() {
+  public @NotNull Thread getThread() {
     return thread;
   }
 }